Add files via upload
In this example, I'm assuming that the solar panel and weather data is stored in a CSV file called 'solar_data.csv'. The data contains columns for temperature, humidity, cloud cover, and solar output. We use the Pandas library to load the data into a DataFrame, and then split the data into training and testing sets using the train_test_split function from Scikit-Learn.

Next, we train a linear regression model on the training data using the LinearRegression class from Scikit-Learn. We use the model to make predictions on the test data, and then calculate the mean squared error of the predictions using the mean_squared_error function from Scikit-Learn.

Of course, this is just a simple example, and there are many ways you could improve this code and model depending on specific needs and data.